Bug 53039 - HTTP Request : Be able to handle responses which size exceeds 2147483647 bytes Fix deprecated calls to setBytes(int) Bugzilla Id: 53039
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1765448 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: a44c6efe5e0730d362415d968d4c904241f327d7

+++ src/protocol/jms/org/apache/jmeter/protocol/jms/sampler/SubscriberSampler.java
@@ -191,7 +191,7 @@ public class SubscriberSampler extends BaseJMSSampler implements Interruptible, 
         if (getReadResponseAsBoolean()) {
             result.setResponseData(buffer.toString().getBytes()); // TODO - charset?
         } else {
-            result.setBytes(buffer.toString().length());
+            result.setBytes((long)buffer.toString().length());
         }
         result.setResponseHeaders(propBuffer.toString());
         if (read == 0) {
